+++ src/simultipac/cst/simulation_results.py
@@ -241,6 +241,68 @@ class CSTResults(SimulationResults):
             **kwargs,
         )
 
+    def fit_alpha(
+        self,
+        fitting_periods: int,
+        running_mean: bool = False,
+        log_fit: bool = True,
+        minimum_final_number_of_electrons: int = 0,
+        bounds: tuple[list[float], list[float]] = (
+            [1e-10, -10.0],
+            [np.inf, 10.0],
+        ),
+        initial_values: list[float] = [0.0, 0.0],
+        minimum_number_of_points: int = 20,
+        min_points_per_period: int = 5,
+        **kwargs,
+    ) -> None:
+        """Fit exp growth factor.
+
+        Parameters
+        ----------
+        fitting_periods :
+            Number of periods over which the exp growth is searched. Longer is
+            better, but you do not want to start the fit before the exp growth
+            starts.
+        running_mean :
+            To tell if you want to average the number of particles over one
+            period. It is recommended with CST, but does not bring anything for
+            SPARK3D. The default is False.
+        log_fit :
+            To perform the fit on :func:`exp_growth_log` rather than
+            :func:`exp_growth`. The default is True, as it generally shows
+            better convergence.
+        minimum_final_number_of_electrons :
+            Under this final number of electrons, we do no bother finding the
+            exp growth factor and set all fit parameters to ``NaN``.
+        bounds :
+            Upper bound and lower bound for the two variables: initial number
+            of electrons, exp growth factor.
+        initial_values: list[float], optional
+            Initial values for the two variables: initial number of electrons,
+            exp growth factor.
+        minimum_number_of_points :
+            Minimum number of fitting points; under this limit, a warning is
+            issued. For CST, should be at least 10 or 20. With SPARK3D, there
+            are two points per RF period so a value of 2 or 4 should be enough.
+        min_points_per_period :
+            Minimum number of points per period. In SPARK3D, we only have two
+            points per RF period so this number should be lower to avoid
+            unnecessary warnings.
+
+        """
+        return super().fit_alpha(
+            fitting_periods=fitting_periods,
+            running_mean=running_mean,
+            log_fit=log_fit,
+            minimum_final_number_of_electrons=minimum_final_number_of_electrons,
+            bounds=bounds,
+            initial_values=initial_values,
+            minimum_number_of_points=minimum_number_of_points,
+            min_points_per_period=min_points_per_period,
+            **kwargs,
+        )
+
 
 class CSTResultsFactory(SimulationResultsFactory):
     """Define an object to easily instantiate :class:`.CSTResults`."""